+++ src/components/home/recipesGrid/constants.ts
@@ -1,0 +1,350 @@
+import { RecipeDetails } from "@/types/Recipe"
+
+export const TEST_ARRAY: RecipeDetails[] = [
+  {
+    category: "Category",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category2",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category2",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category2",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category2",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category2",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category2",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category2",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category2",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category2",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category3",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category3",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category3",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category3",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category3",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category3",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category3",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category3",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category3",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category4",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category4",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category4",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category4",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category4",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category4",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category4",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category4",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category4",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category5",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category5",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category5",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category5",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category5",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category5",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category5",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category5",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category5",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category6",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category6",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category6",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category6",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category6",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category6",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category6",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category6",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category6",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category7",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category7",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+  {
+    category: "Category7",
+    name: "Recipe Name",
+    date: new Date(),
+    imageUrl: "https://bulma.io/images/placeholders/480x640.png",
+  },
+]
+
+export const RECIPE_CATEGORIES = ["All", "Cakes", "Bread", "Cookies"]
+
+export const MAX_PAGE_BUTTONS = 5
