+++ src/types/Recipe.ts
@@ -1,0 +1,6 @@
+export type RecipeDetails = {
+  category: string
+  name: string
+  date: Date
+  imageUrl: string
+}
